Player sorting
Simplified the template by ditchin the "active" and "inactive" sections and just implementing a sorting mechanism that works on all the player records. This is in-memory, and doesn't touch the source data or modify the firebase query at all, but vue recognizes that the object array has changed and updates the view. Seems to work great.

Also got rid of the "Help" route and made it a (nascent) "About" page.